pyproject.toml
[project]
name = "proj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.1",
    "torch>=2.0.0",
    "vllm>=0.19.0",
]
